wrlab: Skip pkg preinst operations in the cross environment
The rpm pkg preinst operations only run in the target.

+++ recipes-connectivity/inetutils/inetutils_1.9.%.bbappend
@@ -2,16 +2,19 @@ RPROVIDES_${PN}-telnet = "telnet"
 RPROVIDES_${PN}-rsh = "rsh"
 
 pkg_preinst_${PN}-doc () {
-    if [ -f ${mandir}/man1/hostname.1 ]; then
-        rm -f ${mandir}/man1/hostname.1
-    fi
-    if [ -f ${mandir}/man1/dnsdomainname.1 ]; then
-        rm -f ${mandir}/man1/dnsdomainname.1
-    fi
-    if [ -f ${mandir}/man1/logger.1 ]; then
-        rm -f ${mandir}/man1/logger.1
-    fi
-    if [ -f ${mandir}/man8/syslogd.8 ]; then
-        rm -f ${mandir}/man8/syslogd.8
+    #!/bin/sh -e
+    if [ x"$D" = "x" ]; then
+        if [ -f ${mandir}/man1/hostname.1 ]; then
+            rm -f ${mandir}/man1/hostname.1
+        fi
+        if [ -f ${mandir}/man1/dnsdomainname.1 ]; then
+            rm -f ${mandir}/man1/dnsdomainname.1
+        fi
+        if [ -f ${mandir}/man1/logger.1 ]; then
+            rm -f ${mandir}/man1/logger.1
+        fi
+        if [ -f ${mandir}/man8/syslogd.8 ]; then
+            rm -f ${mandir}/man8/syslogd.8
+        fi
     fi
 }
